requirements.txt
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
plotly
jinja2
pyyaml
tqdm
tenacity
python-dotenv
openai>=1.0,<2
streamlit>=1.28
requests
openpyxl
